requirements.txt
flask
python-dotenv
sqlalchemy
requests
gunicorn
fpdf2
cryptography
beautifulsoup4
lxml
flask-login
werkzeug
flask-mail
flask-limiter
redis
pytest
pytest-cov
pytest-mock
flask-apscheduler
psycopg2-binary
alembic
prometheus-flask-exporter
flasgger